refactor(Problem6,7): 클래스 분리
