+++ src/main/java/onboarding/problem7/Constant.java
@@ -1,0 +1,7 @@
+package onboarding.problem7;
+
+public class Constant {
+    public static final int FRIEND_WEIGHT = 10;
+    public static final int VISITOR_WEIGHT = 1;
+    public static final int MAX_RECOMMEND = 5;
+}
