+++ src/main/java/onboarding/problem7/Validation.java
@@ -1,0 +1,4 @@
+package onboarding.problem7;
+
+public class Validation {
+}
